+++ alembic/env.py
@@ -1,11 +1,10 @@
 import os
 from logging.config import fileConfig
 
+from alembic import context
 from dotenv import load_dotenv
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
-
-from alembic import context
 
 load_dotenv()
 
@@ -30,7 +29,12 @@ if config.config_file_name is not None:
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-from app.db.models import Base, Audit, User, Note, RateLimit
+try:
+    from app.db.models import Base, Audit, User, Note, RateLimit
+
+    print("Importazione riuscita.")
+except ImportError as e:
+    print(f"Errore di importazione: {e}")
 target_metadata = Base.metadata
