+++ app/core/__init__.py
@@ -1,15 +1,14 @@
 """
 This module contains the core functionality of the application.
 """
-from .access_token import *
-from .rate_limit import *
-from .settings import *
+from app.core.middleware import *
+from .config import *
+from .security import *
 
 __all__ = [
     "settings",
     "create_access_token",
     "decode_access_token",
     "generate_user_token_and_return_user",
-    "RateLimitMiddleware",
     "generate_user_token"
 ]
