+++ app/core/exeptions/generic.py
