+++ app/core/exeptions/note.py
