+++ app/core/middleware/middleware.py
@@ -7,8 +7,8 @@ from fastapi import Request, HTTPException
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.responses import JSONResponse
 
-from app.core.access_token import decode_access_token
-from app.core.settings import settings
+from app.core.config import settings
+from app.core.security import decode_access_token
 from app.db.models.rate_limit import RateLimit
