+++ app/core/security.py
@@ -2,11 +2,11 @@
 Access Token
 """
 from datetime import datetime, timedelta
-import jwt
 
+import jwt
 from fastapi import HTTPException
 
-from app.core.settings import settings
+from app.core.config import settings
 
 
 def create_access_token(data: dict, expires_delta: timedelta | None = None) -> str:
