+++ app/schemas/base.py
@@ -1,20 +1,11 @@
-"""
- PaginatedResponse
-"""
-from typing import Generic, TypeVar, List
+from typing import Generic, List, Optional, TypeVar
+
 from pydantic import BaseModel
 
 T = TypeVar('T')
 
+
 class PaginatedResponse(BaseModel, Generic[T]):
-    """
-    PaginatedResponse
-    items: List[T]
-    total: int
-    page: int
-    page_size: int
-    total_pages: int
-    """
     items: List[T]
     total: int
     page: int
@@ -22,4 +13,4 @@ class PaginatedResponse(BaseModel, Generic[T]):
     total_pages: int
     has_next: bool
     has_prev: bool
-    search_query: str
+    search_query: Optional[str] = None
