+++ app/services/auth/login/repository.py
@@ -1,11 +1,11 @@
 """
 Session actions
 """
-from app.core.access_token import generate_user_token_and_return_user
+from app.core.exeptions.auth import AuthErrorHandler
+from app.core.security import generate_user_token_and_return_user
 from app.db.models.users import User
-from app.services.audit.actions import log_audit_event
-from app.services.error.auth import AuthErrorHandler
-from app.services.logger.actions import LoggerService
+from app.services.audit.repository import log_audit_event
+from app.services.logger.repository import LoggerService
 
 logger = LoggerService().logger
