+++ app/services/user/repository.py
@@ -8,12 +8,12 @@ from fastapi import HTTPException
 from pydantic.v1 import EmailStr
 from sqlalchemy import or_
 
-from app.core.access_token import generate_user_token_and_return_user, decode_access_token
+from app.core.exeptions.user import UserErrorHandler
+from app.core.security import generate_user_token_and_return_user, decode_access_token
 from app.db.models.users import User
 from app.email.email_service import EmailService, EmailSchema
-from app.services.audit.actions import log_audit_event
-from app.services.error.user import UserErrorHandler
-from app.services.logger.actions import LoggerService
+from app.services.audit.repository import log_audit_event
+from app.services.logger.repository import LoggerService
 
 logger = LoggerService().logger
